requirements.txt
anthropic
fastapi
uvicorn[standard]
python-multipart
gspread
google-auth
twilio
